Add files via upload
LRU stands for Least Recursive Used algorithm. This technique is used in cache replacement techniques. 
When new data is hit, this algorithm looks for old data and that is replaced by new data.

Refer this YouTube video for complete details:
https://www.youtube.com/watch?v=akFRa58Svug&t=395s

I have coded as per the explanation made in this video.